Refactor items, quests and hideout modules.
Maintenance of the repo became impossible with so many lines in the items.json. We're going to move toward separate files for each to help fix this issue.

If you relied on monolithic files, it should be very easy to merge all the files back together for your use cases.